jednak wraca DAO i Service, nie zauwazylem, ze takie sa wymagania projektu :(

+++ src/main/java/com/sklep/model/User.java
@@ -13,7 +13,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
 
     @Column(nullable = false, unique = true, length = 50)
     private String username;
@@ -38,8 +38,8 @@ public class User {
     }
 
     // Gettery i settery
-    public int getId() { return id; }
-    public void setId(int id) { this.id = id; }
+    public long getId() { return id; }
+    public void setId(long id) { this.id = id; }
 
     public String getUsername() { return username; }
     public void setUsername(String username) { this.username = username; }
